Perform a sanity check, hopefully fixed issue with writing to a file.
