+++ compare_results.py
@@ -51,22 +51,26 @@ if __name__ == '__main__':
 	good_times = extract_times_from_csv("shcu_good_data.csv", "timestamp_utc")
 	missing_images, empty_images, missing_masks, empty_masks = why_bad_images(good_times)
 
+	for e in missing_masks:
+		print(type(e))  # Just a sanity check. Expecting 'str'
+		break
+
 	print("Writing to missing_images.txt. There are {} missing images".format(missing_images))
-	with open('missing_images.txt', 'wb') as file:
+	with open('missing_images.txt', 'w') as file:
 		for time in missing_images:
 			file.write(time + '\n')
 
 	print("Writing to empty_images.txt. There are {} empty images".format(empty_images))
-	with open('empty_images.txt', 'wb') as file:
+	with open('empty_images.txt', 'w') as file:
 		for time in empty_images:
 			file.write(time + '\n')
 
 	print("Writing to 'missing_masks.txt'. There are {} missing masks.".format(len(missing_masks)))
-	with open('missing_masks.txt', 'wb') as file:
+	with open('missing_masks.txt', 'w') as file:
 		for time in missing_masks:
 			file.write(time + '\n')
 
 	print("Writing to 'empty_masks.txt'. There are {} empty masks.".format(len(empty_masks)))
-	with open('empty_masks.txt', 'wb') as file:
+	with open('empty_masks.txt', 'w') as file:
 		for time in empty_masks:
 			file.write(time + '\n')
